+++ lib/SomfyProtocol/SomfyProtocol.h
@@ -16,6 +16,6 @@
 void sendCommand(byte command, Controller *c);
 void sendCustomCommand(byte sync);
 void buildFrame(byte command, Controller* c);
-void buildCustomFrame(byte command, int remoteId, unsigned long rollingCode);
+void buildCustomFrame(byte command, u_int32_t remoteId, u_int32_t rollingCode);
 
 #endif
